Added original game code
For reference, added the original BASIC game code. I also added instructions for the game, and my notes on the various variables used in the BASIC code. These notes are helpful for translating the BASIC to C#.

Note that the original BASIC code has been reformatted to break up lines containing multiple commands to make it easier to follow, plus a ton of commenting thrown in by myself. Code which has not yet been fully converted to C# is indented 1 tab stop, and the plan is to unindent it once it is converted.